Rename package Remove unnecessary circleci environment variable Remove dependencies we dont need
