+++ pyproject.toml
@@ -1,5 +1,5 @@
 [tool.poetry]
-name = "perf_tools"
+name = "chti"
 version = "0.1.0"
 description = ""
 authors = ["Sébastien Eustace <sebastien@eustace.io>"]
@@ -7,25 +7,17 @@ authors = ["Sébastien Eustace <sebastien@eustace.io>"]
 [tool.poetry.dependencies]
 python = "^3.7"
 plac = "^1.1"
-geojson = "^2.5.0"
-shapely = "^1.7.0"
-folium = "^0.10.1"
 scikit-learn = "^0.22.2"
-mlrose = "^1.3.0"
 jsonpickle = "^1.3"
-aiohttp-index = "^0.1"
 matplotlib = "^3.2.1"
 colorhash = "^1.0.2"
-jsonpath-ng = "^1.5.1"
 pandas = "^1.1.2"
-dask = "^2.26.0"
 numpy = "^1.19.2"
+dask = "^2.26.0"
 toolz = "^0.10.0"
-tornado = "^6.0.4"
 fsspec = "^0.8.2"
 
 [tool.poetry.dev-dependencies]
 black = { version = "*", allow-prereleases = true }
 pytest = "^5.3"
 flake8 = "^3.7"
-pytest-asyncio = "^0.10.0"
